Add extra exception tests for Cortex-R52
I had to move the DFSR read to after the alignment check was disabled, otherwise it double-faulted.

Also it turns out the versatileab tests were running with the Cortex-A linker script not the Cortex-R linker script, and the Cortex-R linker script had a bug in it that we previously missed (it wasn't calling the new exception handlers).