Re-organised handler docs, and renamed the handlers.
* AbortHandler is now DataAbort
* PrefetchHandler is now PrefetchAbort
* UndefinedHandler is now Undefined
* SvcHandler is now SupervisorCall

I also renamed the asm handlers for abort and prefetch to data_abort and prefetch_abort, which makes it clearer that both are aborts and both are handled in abort mode. I also added links to the relevant part of the Arm Armv7 ARM.

This patch also causes `#[exception(Nonsense)` to correctly identify 'Nonense' as the thing that is wrong, rather than the function that comes after the attribute.